Now that we have rate limits under control, we should always try a HEAD request on the indexes & manifests before the PUT request (since the PUT is much slower).

Some local testing for `riscv64`, `arm32v6`, and `s390x` using https://github.com/docker-library/meta/commit/6c3188a828643e981b2d1c37cf3851ed9ec2b4b0 shows some significant speed increases. I chose these arches since they were fully pushed and would only do head requests (and my Docker Hub user doesn't have access to push to the arch namespaces anyway).
```console
$ jq -L ../meta-scripts/ 'include "deploy"; arch_tagged_manifests("riscv64") | deploy_objects[]' builds.json > deploy.json
$ time ../meta-scripts/bin/deploy < deploy.json
...
real    0m6.736s
user    0m0.065s
sys     0m0.032s
$ # compared to Jenkins: ~ 2min 49s - 4min 49s

$ jq -L ../meta-scripts/ 'include "deploy"; arch_tagged_manifests("arm32v6") | deploy_objects[]' builds.json > deploy.json
$ time ../meta-scripts/bin/deploy < deploy.json
...
real    0m32.444s
user    0m0.053s
sys     0m0.035s
$ # compared to Jenkins: ~ 5min 16s - 8min 58s

$ jq -L ../meta-scripts/ 'include "deploy"; arch_tagged_manifests("s390x") | deploy_objects[]' builds.json > deploy.json
$ time ../meta-scripts/bin/deploy < deploy.json
...
real    3m12.348s
user    0m0.278s
sys     0m0.252s
$ # compared to Jenkins: ~ 19min 36s - 24min 22s
```