@yosifkit 2024-06-25
> So this only removes the size check for manifests? Are all our requests still properly accounted for when considering pull and abuse limits?

Yup, the rate limiting is active on every request to Docker Hub using this `registry` client (https://github.com/docker-library/meta-scripts/blob/9c6f9d47a24ff6940ee3071e60f2456669f16b23/registry/client.go#L35)