@tianon 2024-06-25
Because of the way that's implemented, it technically applies to requests to `auth.docker.io` too (any HTTP request that's in the path to our actual lookups gets accounted individually, even including redirects), so it's _technically_ a little tiny bit overaggressive, but I think it's probably better to limit too much than too little. :sweat_smile: