@tianon 2024-06-25
Confirmed, `deploy-riscv64` went from ~3m15s down to ~13s :joy: